backend/spark/consumer/requirements.txt
kafka-python==2.0.2
pyspark==3.5.0
mysql-connector-python==8.2.0
redis==5.0.1
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
